requirements.txt
fastapi~=0.110
uvicorn[standard]~=0.29
pydantic~=1.10
python-dotenv~=1.0
httpx~=0.27
requests~=2.31
openai>=1.14,<2.0
